make sure to flush the event cache before playing any song, otherwise events from last somg might still pollute us
git-svn-id: https://svn.code.sf.net/p/dosmid/code/trunk@81 850bf26b-01f9-4b63-b2ae-8b26f6eb5263

+++ midi.c
@@ -381,7 +381,7 @@ long midi_track2events(FILE *fd, char **title, int titlenodes, int titlemaxlen, 
  * to the unique track. I take care here to not allocate/free memory here.
  * All notes are already in RAM after all. totlen is filled with the total
  * time of the merged tracks (in miliseconds). */
-long midi_mergetrack(long t0, long t1, unsigned long *totlen, unsigned int timeunitdiv) {
+long midi_mergetrack(long t0, long t1, unsigned long *totlen, unsigned short timeunitdiv) {
   long res = -1, lasteventid = -1, selectedid;
   int selected;
   unsigned long curtempo = 500000l, utotlen = 0;
